+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## Changelog for v4.0 release
+
+Changes:
+
+- Added support for user-editable configuration file (i.e. `infiren.conf`)
+- Added feature to load/save command history from/to file on startup/exit
+- Added feature to save/load macros to/from file (commands `save-macro`/`load-macro`)
+- Extended `undo` command to allow undoing/redoing entire macros
+- Reworked several commands to accomodate new commands (e.g. `start-macro` + `end-macro` -> `record-macro`)
+- Applied various minor code modification (variables, comments, errors, etc.)
+
 ## Changelog for v3.11 release
 
 Initial release (versions prior to v3.11 have not been published).
@@ -12,4 +23,4 @@ Features:
 
 ##
 
-_Last updated: 09/13/23_
+_Last updated: 09/25/23_